+++ index.html
@@ -235,6 +235,6 @@
 </body>
 </html>
 <head>
-  <script src = "./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </head>
