+++ js/data.js
@@ -1,0 +1,28 @@
+const PHOTOS_COUNT = 25;
+const COUNT_COMMENTS=10;
+const DESCRIPTIONS = [
+  'Ура, скоро Новый Год!!!:)',
+  'Вчера ходила в кино... лучше бы не ходила',
+  'Я больше не буду ходить без шапки',
+  'Живите, кайфуйте, жизнь одна',
+  'Не баг, а фича;)'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Виктория',
+  'Кристина',
+  'Роман',
+  'Анлегина',
+  'Андрей',
+];
+
+const photos=[];
